# Replace buffer with bufferedReader